+++ demos/video/experiment.py
@@ -5,6 +5,8 @@ from psynet.media import prepare_s3_bucket_for_presigned_urls
 from psynet.modular_page import ModularPage, VideoPrompt, VideoRecordControl
 from psynet.page import SuccessfulEndPage, UnsuccessfulEndPage
 from psynet.timeline import (
+    Event,
+    MediaSpec,
     PageMaker,
     PreDeployRoutine,
     ProgressDisplay,
@@ -24,6 +26,50 @@ video_record_page = join(
         "prepare_s3_bucket_for_presigned_urls",
         prepare_s3_bucket_for_presigned_urls,
         {"bucket_name": bucket_name, "public_read": True, "create_new_bucket": True},
+    ),
+    ModularPage(
+        "simple_video_prompt",
+        VideoPrompt(
+            "/static/flower.mp4",
+            flask.Markup(
+                """
+            <h3>Example video prompt:</h3>
+            <p><a href="https://commons.wikimedia.org/wiki/File:Water_lily_opening_bloom_20fps.ogv">SecretDisc</a>, <a href="https://creativecommons.org/licenses/by-sa/3.0">CC BY-SA 3.0</a>, via Wikimedia Commons</p>
+            """
+            ),
+        ),
+        time_estimate=5,
+    ),
+    ModularPage(
+        "video_play_window",
+        VideoPrompt(
+            "/static/flower.mp4",
+            flask.Markup(
+                """
+            <h3>Example video prompt with play window:</h3>
+            <p><a href="https://commons.wikimedia.org/wiki/File:Water_lily_opening_bloom_20fps.ogv">SecretDisc</a>, <a href="https://creativecommons.org/licenses/by-sa/3.0">CC BY-SA 3.0</a>, via Wikimedia Commons</p>
+            """
+            ),
+            play_window=[3, 4],
+        ),
+        time_estimate=5,
+    ),
+    ModularPage(
+        "video_plus_audio",
+        VideoPrompt(
+            "/static/birds.mp4",
+            "Here we play a video, muted, alongside an audio file.",
+            muted=True,
+        ),
+        time_estimate=5,
+        media=MediaSpec(audio={"soundtrack": "/static/funk-game-loop.mp3"}),
+        events={
+            "playSoundtrack": Event(
+                is_triggered_by="promptStart",
+                delay=0.0,
+                js="psynet.audio.soundtrack.play()",
+            )
+        },
     ),
     ModularPage(
         "video_record_page",